refactor: change calendar page to scheduçe
