+++ src/pages/schedule/schedule.style.scss
@@ -1,0 +1,48 @@
+@use "../../style/variables" as var;
+@use "../../style/mixins.scss" as mixin;
+
+.schedule {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
+  width: 100%;
+  gap: 2rem;
+
+  &__warning {
+    position: absolute;
+  }
+
+  &__title {
+    color: var.$white;
+    text-align: center;
+    font-size: 24px;
+    line-height: 100%;
+  }
+
+  &__button {
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    gap: 6px;
+    font-weight: 500;
+    text-align: center;
+    background-color: var.$blue;
+    color: var.$blue-white;
+    font-size: 12px;
+    border: none;
+    border-radius: 8px;
+    padding: 8px 12px;
+
+    @include mixin.desktopStyle() {
+      font-size: 16px;
+      padding: 8px 16px;
+
+      &__icon {
+        width: 16px;
+        height: 16px;
+      }
+    }
+  }
+}
